Split out bokeh_dof and copy effects
