+++ servers/rendering/renderer_rd/shaders/effects/copy_to_fb.glsl
